refactor: create model.js, view.js to seperate different function from original files
